requirements.txt
black
pytest
openai
langchain
langchain-openai
langchain-community
langchain-core
click
streamlit
ipython
pylint
pytest
python-dotenv
fire
git+https://github.com/openai/whisper.git 